adjust folder structure
