+++ plasmic.json
@@ -31,7 +31,7 @@
           "projectId": "mTVXT6w3HHjZ4d74q3gB76",
           "renderModuleFilePath": "plasmic/flip_card/PlasmicUpdate.tsx",
           "importSpec": {
-            "modulePath": "Update.tsx"
+            "modulePath": "update/index.tsx"
           },
           "cssFilePath": "plasmic/flip_card/PlasmicUpdate.css",
           "scheme": "blackbox"
@@ -43,7 +43,7 @@
           "projectId": "mTVXT6w3HHjZ4d74q3gB76",
           "renderModuleFilePath": "plasmic/flip_card/PlasmicMetric.tsx",
           "importSpec": {
-            "modulePath": "Metric.tsx"
+            "modulePath": "metric/index.tsx"
           },
           "cssFilePath": "plasmic/flip_card/PlasmicMetric.css",
           "scheme": "blackbox"
@@ -55,7 +55,7 @@
           "projectId": "mTVXT6w3HHjZ4d74q3gB76",
           "renderModuleFilePath": "plasmic/flip_card/PlasmicIconButton.tsx",
           "importSpec": {
-            "modulePath": "IconButton.tsx"
+            "modulePath": "iconButton/index.tsx"
           },
           "cssFilePath": "plasmic/flip_card/PlasmicIconButton.css",
           "scheme": "blackbox"
