+++ src/component/plasmic/flip_card/PlasmicUpdate.tsx
@@ -27,8 +27,8 @@ import {
   StrictProps,
   deriveRenderOpts
 } from "@plasmicapp/react-web";
-import Metric from "../../Metric"; // plasmic-import: _i6uD1XPzdbux6R/component
-import IconButton from "../../IconButton"; // plasmic-import: odPjbfT2kyJgB_S/component
+import Metric from "../../metric/index"; // plasmic-import: _i6uD1XPzdbux6R/component
+import IconButton from "../../iconButton/index"; // plasmic-import: odPjbfT2kyJgB_S/component
 
 import "@plasmicapp/react-web/lib/plasmic.css";
 import "../plasmic__default_style.css"; // plasmic-import: global/defaultcss
